+++ libs/iso-bmff/src/createIsoBoxReadableStream.ts
@@ -15,6 +15,6 @@ import type { IsoBoxWriteViewConfig } from './IsoBoxWriteViewConfig.ts'
  *
  * @public
  */
-export function createIsoBoxReadableStream(boxes: Iterable<IsoBoxStreamable>, config: IsoBoxWriteViewConfig = {}): ReadableStream<Uint8Array> {
+export function createIsoBoxReadableStream(boxes: Iterable<IsoBoxStreamable>, config: IsoBoxWriteViewConfig): ReadableStream<Uint8Array> {
 	return new IsoBoxReadableStream(boxes, config)
 }